fix: Parse error for openid expires_in with newer synapse

+++ lib/src/model/open_id_credentials.dart
@@ -31,7 +31,7 @@ class OpenIdCredentials {
       : accessToken = json['access_token'],
         tokenType = json['token_type'],
         matrixServerName = json['matrix_server_name'],
-        expiresIn = json['expires_in'];
+        expiresIn = json['expires_in'].toDouble();
 
   Map<String, dynamic> toJson() {
     final data = <String, dynamic>{};
